requirements.txt
asgiref==3.7.2
autopep8==2.3.1
contourpy==1.3.1
cycler==0.12.1
Django==5.0.9
django-cors-headers==4.6.0
djangorestframework==3.15.2
fonttools==4.55.3
kiwisolver==1.4.7
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.0.0
pycodestyle==2.12.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.9.0
